Add concurrency, sorting, and some clarity
Added in some concurrency so the requests can be sent and received
quicker. Time tests on my machine took execution time for the large
input file from ~550ms to ~275ms, cutting execution time in half. I
suspect this is simply because there are two "machines", Acme and
Madrid, and they are each able to take one at a time. In a more
real-to-life situation I suspect this change might make the code even
faster.

Also added in a sort for the results. Before, the results were printed
in the order they were in the input file. After adding concurrency, they
were returned in a somewhat random order. Thus I sorted them by program
id so they would be easier to find. I may try to find a way to maintain
the input order in the future.